# internal/server/handlers/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gateon Management API",
    "description": "API for managing routes, services, entrypoints, middlewares, TLS options, and global configuration. Requires authentication (PASETO) when RBAC is enabled.\n\n## RBAC\n\n| Role | Read | Write config (routes, services, etc.) | Write users/global |\n|------|------|----------------------------------------|--------------------|\n| admin | ✓ | ✓ | ✓ |\n| operator | ✓ | ✓ | ✗ |\n| viewer | ✓ | ✗ | ✗ |",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "Gateway API" }],
  "paths": {
    "/v1/routes": {
      "get": { "summary": "List routes", "tags": ["routes"], "responses": { "200": { "description": "OK" } } },
      "put": { "summary": "Create or update route (operator, admin)", "tags": ["routes"], "responses": { "200": { "description": "OK" }, "403": { "description": "Insufficient permissions" } } }
    },
    "/v1/routes/{id}": {
      "delete": { "summary": "Delete route (operator, admin)", "tags": ["routes"], "responses": { "204": { "description": "No Content" }, "403": { "description": "Insufficient permissions" } } }
    },
    "/v1/services": {
      "get": { "summary": "List services", "tags": ["services"], "responses": { "200": { "description": "OK" } } },
      "put": { "summary": "Create or update service (operator, admin)", "tags": ["services"], "responses": { "200": { "description": "OK" }, "403": { "description": "Insufficient permissions" } } }
    },
    "/v1/services/{id}": {
      "delete": { "summary": "Delete service (operator, admin)", "tags": ["services"], "responses": { "204": { "description": "No Content" }, "403": { "description": "Insufficient permissions" } } }
    },
...
    "/v1/tls-options": {
      "get": { "summary": "List TLS options", "tags": ["tls"], "responses": { "200": { "description": "OK" } } },
      "put": { "summary": "Create or update TLS option (operator, admin)", "tags": ["tls"], "responses": { "200": { "description": "OK" }, "403": { "description": "Insufficient permissions" } } }
    },
    "/v1/tls-options/{id}": {
      "delete": { "summary": "Delete TLS option (operator, admin)", "tags": ["tls"], "responses": { "204": { "description": "No Content" }, "403": { "description": "Insufficient permissions" } } }
    },
...
  }
}